First commit!
clj-vlc aims to be a simple wrapper to use the fantastic VLC Media
Player (thanks to all those volunteers that created such a great
product!).

I based the communication with VLC on TCP through the useful clj-tcp,
although this may change later.

The interface is easy, start-vlc-server to get a channel, which you
can use to call run-vlc-cmds and available-commands. More info in the
README.org (hopefully all the Org-mode syntax is readable {thanks a
lot Emacs+Org mode people!!}).

Enjoy!